Change file name type to types
