+++ backend/src/game/game.controller.ts
@@ -1,5 +1,5 @@
 import type { FastifyReply, FastifyRequest } from "fastify";
-import { CreateGameDTO, GameIdDTO } from "./game.type.ts";
+import { CreateGameDTO, GameIdDTO } from "./game.types.ts";
 
 export const createGameHandler = async (
     { body }: { body: CreateGameDTO },
@@ -9,9 +9,11 @@ export const createGameHandler = async (
     try {
         //const game = await req.server.gameService.create(body);
         const game = body; // TODO
+
         if (!game) {
             return reply.code(400).send({ error: "Failed to create game" });
         }
+
         return reply.code(201).send(game);
     } catch (error) {
         req.log.error({ err: error }, "Failed to create game");
@@ -26,9 +28,11 @@ export const getGameByIdHandler = async (
 ) => {
     try {
         const game = await req.server.gameService.findById(params.id);
+
         if (!game) {
             return reply.code(404).send({ error: "Game not found" });
         }
+
         return reply.send(game);
     } catch (error) {
         req.log.error({ err: error }, "Failed to get game by ID");
