+++ backend/src/game/game.routes.ts
@@ -1,5 +1,5 @@
 import type { FastifyPluginAsync } from "fastify";
-import { createGameSchema, gameIdSchema } from "./game.type.ts";
+import { createGameSchema, gameIdSchema } from "./game.types.ts";
 import { createGameHandler, getAllGamesHandler, getGameByIdHandler } from "./game.controller.ts";
 import { withZod } from "../utils/zod-validate.ts";
