+++ backend/src/game/game.service.ts
@@ -1,6 +1,6 @@
 import type { FastifyInstance } from "fastify";
 import { games } from "../db/db.schema.ts";
-import { NewGame, Game } from "./game.type.ts";
+import { NewGame, Game } from "./game.types.ts";
 import { eq } from "drizzle-orm";
 
 export const createGameService = (fastify: FastifyInstance) => {
