+++ backend/src/game/game.types.ts
@@ -1,5 +1,9 @@
 import { WebSocket } from "ws";
 import GameEngine from "./game.engine";
+import { z } from "zod";
+import { games } from "../db/db.schema.ts";
+export const createGameSchema = z.object({});
+export const gameIdSchema = z.object({ id: z.coerce.number().int().gt(0) });
 
 export type Player = {
     socket: WebSocket;
@@ -37,3 +41,9 @@ export type GameSession = {
     loop?: NodeJS.Timeout; // Store the interval for the game loop
     engine?: GameEngine; // Store the game engine
 };
+
+export type CreateGameDTO = z.infer<typeof createGameSchema>;
+export type GameIdDTO = z.infer<typeof gameIdSchema>;
+
+export type Game = typeof games.$inferSelect;
+export type NewGame = typeof games.$inferInsert;
